Refocusing
Previously, this project was attempting to create the Service Broker
for the application as well as the application itself.  This might
eventually happen, but it was a bit premature.  This change removes
the Service Broker and focuses on the application itself.

As of this commit, the application has a complete data model,
including REST-ful APIs, documentation, and database schema for
applications, schedules, and chaos linking them together.  There is no
action on that chaos, but at least we can model the data required to
perform chaos.