Instance Destruction
This change added an asynchronous destruction component to the
application.  This component listens for the creation of schedules and
when one occurs, adds it to a queue.  When the schedule activates
(based on its expression), each chaos related to that schedule is
triggered.  Each of these chaoses dictates a probability that an
application's instances will be terminated.  After the chaos is
complete, an event record is creating documenting what instances were
terminated.

Note that due to incomplete APIs, termination does not yet happen
successfully.